More update on disk_manager.rs

+++ src/common/config.rs
@@ -1,5 +1,5 @@
 pub(crate) type FrameId = i32;
 pub(crate) type PageId = i32;
 
-pub(crate) const DEFAULT_DB_IO_SIZE: u64 = 16;
-pub(crate) const BUSTUB_PAGE_SIZE: u64 = 4096;    
+pub(crate) const DEFAULT_DB_IO_SIZE: usize = 16;
+pub(crate) const BUSTUB_PAGE_SIZE: usize = 4096;
